vvtest: automatically soft link test file
Now the test specification file is always soft linked in the TestResults
directory.  You don't have to specifically add the test specification
file into the list of source files to be linked.

+++ libvvtest/tests/testutils.py
@@ -161,7 +161,7 @@ def remove_results():
 def rmallfiles():
     for f in os.listdir("."):
       if f != "conflicts.out":
-        if os.path.isdir(f):
+        if not os.path.islink(f) and os.path.isdir(f):
           shutil.rmtree(f)
         else:
           os.remove(f)
